bug workaround
when flight is enabled and a players gamemode is swapped from a flying gamemode to a non flying gamemode (creative -> survival, spectator -> adventure) after executing /fly on the player the plugin will crash. isFlying check is kinda dodgey as it checks if player is in double tap space flight rather than if the player has the ability to fly.